refactor(cli): extract add feishu|lark onboarding into cli-add-feishu.ts
runAdd had grown ~300 lines of cloud-facing onboarding wiring (the lark
guided-console closure + createFeishuAppFlow + the .env reader), crowding the
command dispatcher. Move it behind one seam — onboardFeishuCloudApp(target,
kind, envIgnored) — in a command rendering layer beside cli-models.ts /
cli-auth.ts, keeping cli.ts to arg parsing and process side effects. The
extraction also picks up the shared registration-error predicates and derives
API origins from the cloud profile instead of repeating literals. No behavior
change.
